# contract/openapi.yaml
openapi: 3.1.0
info:
  title: sirus Core API
  version: 0.4.0
  description: |
    Shared contract for all sirus surfaces (cli, gui, web, auto).

    **Golden rule:** this API is the single source of truth. CLI, GUI, Web, and
    Automation are all clients. None talk to the scan engine or the Cerebus
    guardrail directly.

    This spec covers the endpoints the CLI consumes. Endpoints owned entirely by
    the web/GUI surfaces (teams, integrations, audit-log, SSO) are listed in
    `docs/system-overview.md` but not yet specified here.

    Deviations from `docs/original-prd.md` are recorded in `docs/decisions.md`
    and marked `x-decision` below.
servers:
  - url: https://api.sirus.dev/api/v1
    description: Production
  - url: http://localhost:4010
    description: Prism mock (set SIRUS_API_URL to this)
security:
  - apiKey: []
...
paths:
  /scans:
    post:
      tags: [scans]
      operationId: createScan
      summary: Create a scan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScanCreate' }
      responses:
        '202':
          description: Scan queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Scan' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    get:
      tags: [scans]
      operationId: listScans
      summary: List scans
      parameters:
        - { name: project_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ScanStatus' } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Scan' }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        Project API key for CLI and CI (**K** in the PRD's auth column). Web and
        GUI use a session cookie / JWT (**S**) instead.

        x-decision: D-004 — the WebSocket at `/scans/{id}/stream` takes the same
        `Authorization: Bearer <key>` header on the upgrade request, with a
        `?token=` query fallback for browser clients, which cannot set headers on
        a WebSocket handshake. Auth failure closes with code 4401.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    Problem:
      description: RFC-7807 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
...
    ScanStatus:
      type: string
      enum: [queued, running, completed, failed, canceled]
    ScanSource:
      type: string
      enum: [upload, git, inline]
    ScanTrigger:
      type: string
      enum: [manual, ci, webhook, schedule]
...
    FailOn:
      type: string
      enum: [all, new, verified-secrets]
      description: |
        The gate *predicate*. Deliberately diverges from Snyk's own
        `all|upgradable|patchable`.

        x-decision: D-003 — `fail_on` selects the predicate, `severity_threshold`
        sets the bar. The PRD's ANSI mockup conflates the two by printing
        `gate: fail-on=high`.
...
    Problem:
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri, examples: ['https://sirus.dev/errors/rule-invalid'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Non-standard sixth member, `SIRUS_ERR_*` namespace.
          examples: [SIRUS_ERR_RULE_SCHEMA, SIRUS_ERR_PARSE]
    SeverityCounts:
      type: object
      properties:
        critical: { type: integer }
        high: { type: integer }
        medium: { type: integer }
        low: { type: integer }
        info: { type: integer }
    ScanCreate:
      type: object
      required: [project_id]
      properties:
        project_id: { type: string, format: uuid }
        source: { $ref: '#/components/schemas/ScanSource' }
        git_ref: { type: string, examples: ['refs/pull/42/head'] }
        commit_sha: { type: string }
        baseline_commit: { type: string }
        diff_aware: { type: boolean, default: false }
        rulesets:
          type: array
          items: { type: string }
          examples: [['p/fintech-core', 'p/secrets']]
        policy_id: { type: string, format: uuid }
        validate_secrets: { type: boolean, default: false }
        severity_threshold:
          $ref: '#/components/schemas/Severity'
          description: |
            x-decision: D-002 — added to the PRD's body. The server echoes an
            `exit_code` on `scan.completed`, but the CLI computes its own gate
            verdict locally and treats the server value as a cross-check, so that
            CI behavior stays deterministic offline and under --replay.
        fail_on: { $ref: '#/components/schemas/FailOn' }
    Scan:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/ScanStatus' }
        source: { $ref: '#/components/schemas/ScanSource' }
        trigger: { $ref: '#/components/schemas/ScanTrigger' }
        git_ref: { type: [string, 'null'] }
        commit_sha: { type: [string, 'null'] }
        baseline_commit: { type: [string, 'null'] }
        compliance_score: { type: [number, 'null'], examples: [72.5] }
        money_at_risk_inr:
          type: [integer, 'null']
          description: Scan-level aggregate over all findings, not just rendered ones.
          examples: [5120000]
        counts: { $ref: '#/components/schemas/SeverityCounts' }
        exit_code: { type: [integer, 'null'], enum: [0, 1, 2, 3, null] }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
...